Added changes to player controller so the avg position of the tavern is in the center of the screen.
Made it so the more modules you place the further out your able to zoom.

Made it so id you place a roof on a roof the bottom roof turns into a building module.

Made it so if you place a roof on the side of another roof the new module is a roof.