You can now press enter to change between the tavern builder and tavern shift modes.
There are some strange issues however:

The first time you press the button nothing happens and the second time you press it it transistions to the tavern shift player however it does it weirdly and the view blend glitches out a bit.

any attempts to switch after this seem to work fine however ... Yay?